requirements.txt
adafruit-ampy
click
esptool
pyserial
pytest
requests
requests-mock
setuptools
